backend/requirements.txt
# VARDAx Backend Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0
httpx==0.25.2

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# ML & Data Processing
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.4.0
joblib==1.3.2

# Database (optional for full deployment)
# sqlalchemy==2.0.25
# asyncpg==0.29.0
# redis==5.0.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
python-dateutil==2.8.2
